requirements.txt
pytz
selenium
PyAutoGUI
ipykernel
python-dotenv
python-whatsapp-bot